优化

+++ src/pageDetail/supplierManage/supplierManage.vue
@@ -240,10 +240,25 @@ export default {
       });
     },
     deleteBtn(row) {
-      axios.delete(supplierDelete + "?id=" + row.id).then(data => {
-        this.$message.success("删除成功");
-        this.getList();
+      this.$confirm('此操作将永久删除该厂商, 是否继续?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        axios.delete(supplierDelete + "?id=" + row.id).then(data => {
+          this.getList();
+        });
+        this.$message({
+          type: 'success',
+          message: '删除成功!'
+        });
+      }).catch(() => {
+        this.$message({
+          type: 'info',
+          message: '已取消删除'
+        });
       });
+
     },
     getList() {
       axios.post(supplierSelect, this.theQuery).then(data => {
